Correct logging deps issue
slf4j deps were being included in the agent jar which could
cause really odd runtime failures when running your application
with this agent. The agent included logging deps in the jar,
which could cause issues with your environment. I did not notice this
until running with spring-boot under gradle and a large legacy maven application.
In the interest of preventing future headaches, there is now no log4j or slf4j/logback
dependency included by this agent.

The only non-clojure dependency in the agent is now bytebuddy.
Bytebuddy is a nice library, but I'd like to get rid of it as well.
See https://github.com/tstout/cafebabe if you would like to help.